+++ polilinea.h
@@ -38,13 +38,13 @@ void polilinea_destruir(polilinea_t *polilinea);
 ** Devuelve el extremo en "x" de la polilinea
 ** Si "mayor" es true, devuelve el extremo mayor, sino el menor
 */
-double polilinea_get_extremo_x(polilinea_t *polilinea, bool mayor);
+float polilinea_get_extremo_x(polilinea_t *polilinea, bool mayor);
 
 /*
 ** Devuelve el extremo en "y" de la polilinea
 ** Si "mayor" es true, devuelve el extremo mayor, sino el menor
 */
-double polilinea_get_extremo_y(polilinea_t *polilinea, bool mayor);
+float polilinea_get_extremo_y(polilinea_t *polilinea, bool mayor);
 /*
 ** De ser posible, asigna un punto de coordenadas "x" e "y" en la posición "pos" de la polilinea "polilinea". Caso contrario devuelve false;
 */
